+++ package.json
@@ -9,12 +9,16 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "@fortawesome/fontawesome-svg-core": "^6.4.0",
+    "@fortawesome/free-solid-svg-icons": "^6.4.0",
+    "@fortawesome/react-fontawesome": "^0.2.0",
     "framer-motion": "^10.10.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-icons": "^4.4.0",
+    "react-icons": "^4.8.0",
     "react-router-dom": "^6.10.0",
-    "react-scroll": "^1.8.9"
+    "react-scroll": "^1.8.9",
+    "vite-plugin-react-icons": "^0.2.0"
   },
   "devDependencies": {
     "@types/react": "^18.0.28",
