--- src/data/SkillsData.js
@@ -1,104 +1,0 @@
-import {
-	FaBootstrap,
-	FaCss3Alt,
-	FaGithub,
-	FaHtml5,
-	FaNpm,
-	FaReact,
-	FaSass,
-	FaYarn,
-	FaFigma,
-} from "react-icons/fa";
-import { DiJavascript1 } from "react-icons/di";
-import { VscJson } from "react-icons/vsc";
-import {
-	SiTailwindcss,
-	SiNextdotjs,
-	SiTypescript,
-	SiFirebase,
-	SiStyledcomponents,
-	SiFramer,
-	SiMaterialui,
-	SiRedux
-} from "react-icons/si";
-
-export const SkillsData = [
-	{
-		name: "HTML",
-		// icon: <FaHtml5 />,
-	},
-	{
-		name: "CSS",
-		icon: <FaCss3Alt />,
-	},
-	{
-		name: "Javascript",
-		icon: <DiJavascript1 />,
-	},
-	{
-		name: "TypeScript",
-		icon: <SiTypescript />,
-	},
-	{
-		name: "React Js",
-		icon: <FaReact />,
-	},
-	{
-		name: "Next Js",
-		icon: <SiNextdotjs />,
-	},
-	{
-		name: "Redux",
-		icon: <SiRedux />,
-	},
-	{
-		name: "Firebase",
-		icon: <SiFirebase />,
-	},
-	{
-		name: "JSON",
-		icon: <VscJson />,
-	},
-	{
-		name: "Bootstrap",
-		icon: <FaBootstrap />,
-	},
-	{
-		name: "Tailwind",
-		icon: <SiTailwindcss />,
-	},
-	{
-		name: "Material UI",
-		icon: <SiMaterialui />,
-	},
-	{
-		name: "Sass",
-		icon: <FaSass />,
-	},
-	{
-		name: "Styled Components",
-		icon: <SiStyledcomponents />,
-	},
-	{
-		name: "Framer Motion",
-		icon: <SiFramer />,
-	},
-	{
-		name: "Git",
-		icon: <FaGithub />,
-	},
-	{
-		name: "Npm",
-		icon: <FaNpm />,
-	},
-	{
-		name: "Yarn",
-		icon: <FaYarn />,
-	},
-	{
-		name: "Figma",
-		icon: <FaFigma />,
-	},
-];
-
-export default SkillsData;
