--- src/data/WorkData.js
@@ -1,72 +1,0 @@
-export const WorkData = [
-  {
-    title: "News vile",
-    desc: "This is an application that enables users to read the latest news from various topics and interests around the world. .",
-    tech: [
-      "React.js",
-      "Bootstrap",
-      "MaterialUI",
-      "APIs",
-      "Redux",
-      "Context API",
-    ],
-    gitlink: "//github.com/Maxwell011/capstone-project.git",
-    site: "//capstone-project-g48w4j4mo-maxwell011.vercel.app/",
-  },
-  {
-    title: "Chill Movies",
-    desc: "A Movie website built with TVmaze API that displays movies and TV shows with comments and likes for interactions",
-    tech: ["Javascript", "Css", "APIs", "Redux", "WebPack"],
-    gitlink: "https://github.com/Maxwell011/Chill-Movies.git",
-    site: "//maxwell011.github.io/Chill-Movies/",
-  },
-  {
-    title: "Multicutural Music Summit",
-    desc: "A website for a music festival that displays the artists and their music",
-    tech: ["React.js", "Axios", "APIs", "TailwindCss", "Material UI"],
-    gitlink: "https://github.com/Maxwell011/Music-Conference.git",
-    site: "https://maxwell011.github.io/Music-Conference/",
-  },
-  {
-    title: "Book Store App",
-    desc: "This a Book App that let you add books that you want and let you continue from where you stop.",
-    tech: ["React.js", "Redux", "Javascript", "Css"],
-    gitlink: "https://github.com/Maxwell011/book-store.git",
-    site: "https://book-store-b42ktopqk-maxwell011.vercel.app/",
-  },
-  {
-    title: "A Budget App",
-    desc: "This is a mobile web application designed to help users manage their budget effectively. With this app, users can view their transaction history, grouped by category, to easily track how much money they have spent and on what.",
-    tech: ["Rails", "Ruby Gem", "PostgreSQL", "Bootstrap", "Ruby on Rails"],
-    gitlink: "",
-    site: "//",
-  },
-  {
-    title: "Todo List App",
-    desc: "A to-do list Web application that lets the user add, remove and reorder to do lists and checks a task when completed with a button to delete all completed task",
-    tech: ["Javascript", "Css", ""],
-    gitlink: "https://github.com/Maxwell011/To-Do-list.git",
-    site: "https://maxwell011.github.io/To-Do-list/",
-  },
-  // {
-  //   title: "Onixx Digital",
-  //   desc: "A website for a company that helps with project and site management for the best optimization",
-  //   tech: ["React.js", "Styled-Components"],
-  //   gitlink: "//www.github.com/Pappyjay23/megapod",
-  //   site: "//onixx-digital.netlify.app/",
-  // },
-  // {
-  //   title: "Megapod",
-  //   desc: "A site hosting podcasts",
-  //   tech: ["React.js", "Styled-Components", "Css"],
-  //   gitlink: "//www.github.com/Pappyjay23/megapod",
-  //   site: "//megapod-vp.netlify.app",
-  // },
-  // {
-  //   title: "Lux Fitness",
-  //   desc: "A web app depicting a gym, it's features and plans",
-  //   tech: ["Html", "Css", "JS", "GSAP"],
-  //   gitlink: "//www.github.com/Pappyjay23/Lux-Fitness-Gym",
-  //   site: "//lux-fitness-gym.netlify.app",
-  // },
-];
