+++ src/data/WorkData.jsx
@@ -1,0 +1,51 @@
+export const WorkData = [
+  {
+    title: "News vile",
+    desc: "This is an application that enables users to read the latest news from various topics and interests around the world. .",
+    tech: [
+      "React.js",
+      "Bootstrap",
+      "MaterialUI",
+      "APIs",
+      "Redux",
+      "Context API",
+    ],
+    gitlink: "//www.github.com/Maxwell011/capstone-project",
+    site: "//capstone-project-g48w4j4mo-maxwell011.vercel.app/",
+  },
+  {
+    title: "Chill Movies",
+    desc: "A Movie website built with TVmaze API that displays movies and TV shows with comments and likes for interactions",
+    tech: ["Javascript", "Css", "APIs", "Redux", "WebPack"],
+    gitlink: "//www.github.com/Maxwell011/Chill-Movies",
+    site: "//maxwell011.github.io/Chill-Movies/",
+  },
+  {
+    title: "Multicutural Music Summit",
+    desc: "A website for a music festival that displays the artists and their music",
+    tech: ["React.js", "Axios", "APIs", "Material UI"],
+    gitlink: "//www.github.com/Maxwell011/Music-Conference",
+    site: "//maxwell011.github.io/Music-Conference/",
+  },
+  {
+    title: "Book Store App",
+    desc: "This a Book App that let you add books that you want and let you continue from where you stop.",
+    tech: ["React.js", "Redux", "Javascript", "Css"],
+    gitlink: "//www.github.com/Maxwell011/book-store",
+    site: "//book-store-b42ktopqk-maxwell011.vercel.app/",
+  },
+  {
+    title: "Todo List App",
+    desc: "A to-do list Web application that lets the user add, remove and reorder to do lists and checks a task when completed with a button to delete all completed task",
+    tech: ["Javascript", "Css", ""],
+    gitlink: "//www.github.com/Maxwell011/To-Do-list",
+    site: "//maxwell011.github.io/To-Do-list/",
+  },
+  {
+    title: "A Budget App",
+    desc: "This is a mobile web application designed to help users manage their budget effectively.",
+    tech: ["Rails","PostgreSQL", "Bootstrap", "Ruby on Rails"],
+    gitlink: "",
+    site: "",
+  },
+];
