+++ src/screen/App.jsx
@@ -26,7 +26,7 @@ function App() {
   return loading ? (
     <div className='loader'>
       <div className='svg-wrapper'>
-        <svg height='60' width='320' xmlns='http://www.w3.org/2000/svg'>
+        <svg height='90' width='320' xmlns='http://www.w3.org/2000/svg'>
           <rect className='shape' height='60' width='320' />
         </svg>
         <motion.p
